+++ src/Entry.php
@@ -3,7 +3,7 @@
  * Entry
  *
  * @author    Pronamic <info@pronamic.eu>
- * @copyright 2005-2019 Pronamic
+ * @copyright 2005-2020 Pronamic
  * @license   GPL-3.0-or-later
  * @package   Pronamic\WordPress\Pay\Extensions\GravityForms
  */
@@ -13,7 +13,7 @@ namespace Pronamic\WordPress\Pay\Extensions\GravityForms;
 /**
  * Title: WordPress pay extension Gravity Forms entry
  * Description:
- * Copyright: 2005-2019 Pronamic
+ * Copyright: 2005-2020 Pronamic
  * Company: Pronamic
  *
  * @author  Remco Tolsma